requirements.txt
flask
watchdog
python-dotenv
flask-wtf
email-validator
flask-login
flask-mail
PIL
bcrypt
